Added required cast to int for BMP_SIZE in csv_plotter.py

+++ cockpit/util/csv_plotter.py
@@ -55,7 +55,8 @@ def make_bitmap(hex, text=None):
         dc = wx.MemoryDC()
         dc.SelectObject(bmp)
         w, h = dc.GetTextExtent(text)
-        dc.DrawText(text, (BMP_SIZE[0] - w) / 2,  (BMP_SIZE[1] - h) / 2)
+        dc.DrawText(text, int((BMP_SIZE[0] - w) / 2),
+                    int((BMP_SIZE[1] - h) / 2))
         dc.SelectObject(wx.NullBitmap)
     return bmp
